web_ui: tolerate deleted tickets better
Implement a better algorithm for detecting when tickets don't have
ranks.  Also, don't use the ticket_id as the rank as it may already be
in use via rearranging the backlog.  Instead, choose MAX(rank)+1 as the
new rank.  Make the same change in ticket_created(), since it can suffer
from the same problem.